Fix IPMI on BlueField-2
On some BlueField-2 systems, after powercycling,
the set_emu_param.sh script causes a kernel panic.
This is due to trying to detect/instantiate i2c
devices (ee1004 and jc42) which are only needed
on systems that support external DDRs.
None of the BlueField-2 systems support external
DDRs.
This does not happen after software reset.

Although this is not the source of the
problem, it is important to handle this case.
The source of the issue is the i2c interface
causing a panic. But it will not be handled in
this patch.