Temporary workaround for BlueSphere crash due to ipmb-host
After powercycling, the ipmb-host driver tries to
execute a handshake with the BMC and fails because the
BMC is not up yet. This causes a kernel calltrace and
eventually stops responding because there is and ipmi_msghandler
event thats holding a task for too long (forever).

Just to unblock other teams progress, the workaround for now,
is to delay the loading of the ipmb_host driver by 3mn.

I will get back to this and solve it properly from the drivers.