Support eth0/eth1 fru for infiniband
In switchdev mode, ConnectX interfaces are always named p0/p1.
BlueField in Smart NIC configuration always uses switchdev mode.
The enp*0/1 and ibp*0/1 naming scheme is the default naming
convention of the OS and is used on the BlueWhale.

Verification has been testing FRU 13 and 14 on the BlueWhale
with ib enabled instead of ethernet. So we need to support
parsing ifconfig for "ib*" as well.

RM #2627070