Avoid corruption of ipmb_update_timer file
The customer occasionally observes the corruption of /tmp/ipmb_update_timer
Due to this, the mlx_ipmid.service fails.
I was not able to reproduce this issue in our lab systems and we are not
sure when this happens exactly: at boot time? if someone ran a test or
configuration on a system?
To avoid that, we will hold the ipmb_update_timer value in a variable
instead of the /tmp directory. ipmb_update_timer should be wiped out
after reboot.

The advantage of the previous method, was to allow the customer to read
the ipmb_update_timer from their BMC via ipmitool. This way, they could time
when to read the very large FRUs before they get updated.

RM #2661986