Avoid corruption of ipmb_update_timer file
The customer occasionally observes the corruption of /tmp/ipmb_update_timer
Due to this, the mlx_ipmid.service fails.
I was not able to reproduce this issue in our lab systems and we are not
sure when this happens exactly: at boot time? if someone ran a test or
configuration on a system?
We suspect that this file corruption is a race condition between the
set_emu_param.service and oneshot_emu_param.service which could
try to write/read the file at the same time.

So I removed the oneshot_emu_param.service file and replaced it by
execstartpre.

RM #2661986