Improve logic for fw_info data collection
The fw_info FRU has information about:
1) UEFI, ATF and release version
2) OFED version
3) VPD info
4) connectX fw version and board ID

All data above is collected based on whether
the infiniband/ethernet sysfs entry exists,
which should only be the case for 4).

Correct the logic to always get 1), 2) and 3).
Then 4) is the only one dependent on the sysfs
entry /sys/class/infiniband/mlx5_0/1

RM #2663322