Support old systemd syntax
The "append" keyword used in mlx_ipmid.service and set_emu_param.service
is not supported on the systemd version installed by default on
CentOS8.2. DellEMC would like to keep the default systemd
version so we need to use the "file" keyword instead.

"StandardOutput=append..." syntax only works in systemd 240 and newer.
After a restart of the service, the new logs are appended to the old logs.

"StandardOutput=file..." syntax works in systemd v236 and newer so it is
supported in CentOS8.2.
This limitation here is that the log files contents will be overwritten
each time the service restarts. And nothing will be logged after a service
restart until the log file is deleted.

RM #2796022