Fix missing QSFP EEPROM FRUs
The /tmp directory is not tmpfs on CentOS.
That means some of the supposedly temporary files stored in /tmp
are not deleted after a software reset and the logic to detect
link change could be erroneous. If someone removes a QSFP cable after
SW_RESET, this change will not take effect.
To fix this, write all the temporary files into the /run/emu_param
directory. The /run directory is tmpfs for all OS's.

RM #2935116