Fix emmc_info size
The customer sees that although the emmc_info size matches the
emmc_info_filelen on the BF, the fru inventory of that FRU is
2 bytes off.
This is because the "use%", "available", "used" are all
susceptible to change during run time. So take that into account
and cap the emmc_info file size to an extra 500 characters.

RM #3010116