Fix the out of disk space caused by mlx_ipmid service log
The logrotate for mlx_ipmid log is working correctly. The size of the
log will go back to the original size after rotated.
This is because even though we truncate the file, the ipmi_sim writing
to the file will continue writing at whichever offset it were at last.
So we get a mlx_ipmid log with NULL-bytes up to the point where ipmi_sim
truncated it plus the new entries written to the log.
To fix this, we retart the mlx_ipmid service after rotation, so that the
pointer to the log file can be set correctly.