Fix the out of disk space caused by mlx_ipmid service log
The logrotate for mlx_ipmid log is working correctly. The size of the
log will go back to the original size after rotated.
This is because even though we truncate the file, the ipmi_sim writing
to the file will continue writing at whichever offset it were at last.
So we get a mlx_ipmid log with NULL-bytes up to the point where ipmi_sim
truncated it plus the new entries written to the log.
To fix this, the log file should be opened in O_APPEND mode. So that the
NULL bytes on the top of the log will be ignored, and new entries are
written to the start of the file. Append mode has been introduced in
systemd version 240. We can use that feature to make the log to be
rotated correctly.