Remove dynamic allocation of frus in the mlx-bf.emu
Dynamic allocations of the FRUs in the mlx-bf.emu are not necessary
and in fact are useless because once the mlx_ipmid.service is started
it will fetch data from the mlx-bf.emu only once at start time, then
never again. So even if mlx-bf.emu is modified later, it doesn't affect
mlx_ipmid.service until it is restarted.
Instead just set the max possible size for each FRU in the mlx-bf.emu
file directly.
This solves 2 issues:
1) reduces processing time in set_emu_param.sh
2) avoid a bug related to wiping the data in mlx-bf.emu which is due
   to repopulating the file periodically.

RM #3173561