Fix mlx_ipmid and set_emu_paramm services not active
When search the ipmb ko file in set_emu_param script, the starting-point
of the find command shouldn't be set as root dictionary. This is because
the root dictionary could have some redirect folder like
"/mlnx/autodirect/" that is very large and stored in the remote system.

We should use explicit starting-point of the find command to search.

RM #3236339