Fix IPMB host crash
This commit revert the handler for loading ipmb driver, this is because
the driver loading and retry checking happened at the same time could
cause the driver to crash because the ipmb response mismatch to the
ipmb_host driver.

It also addresses the crash issue by modifying the method to verify
the IPMB host driver's connection to the BMC. The previous
implementation used ipmitool mc info to check the connection, which
could cause the IPMB host driver crash and has been replaced with a
directory existence check for the BMC device under sysfs.